Move use case to proper package
